job-list: abstract idsync specific code
Problem: idsync code is not abstracted very well throughout
job-list.  The context is grabbed and used in several places.

Solution: Abstract away idsync lookup and wait functions into
new functions in idsync.[ch].  Remove public functions
idsync_data_destroy_wrapper() and idsync_data_create(), as they
are no longer needed.
